# backend-api/docs/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "High Street Gym API",
    "description": "JSON REST API for tracking animal sightings"
  },
  "servers": [
    {
      "url": "http://localhost:8081"
    }
  ],
  "basePath": "",
  "paths": {
...
    "/classes": {
      "get": {
        "summary": "get all classes",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "summary": "create a new class",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "datetime": {
                    "example": "any"
                  },
                  "roomID": {
                    "example": "any"
                  },
                  "activityID": {
                    "example": "any"
                  },
                  "trainerID": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/classes/details": {
      "get": {
        "summary": "get classes with all details",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/classes/{id}": {
      "get": {
        "summary": "get a class by id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "summary": "delete a class by id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/classes/": {
      "put": {
        "summary": "update a class by id",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "example": "any"
                  },
                  "datetime": {
                    "example": "any"
                  },
                  "roomID": {
                    "example": "any"
                  },
                  "activityID": {
                    "example": "any"
                  },
                  "trainerID": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/bookings": {
      "get": {
        "summary": "get all bookings",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "summary": "create a new booking",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "409": {
            "description": "Conflict"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userID": {
                    "example": "any"
                  },
                  "classID": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/bookings-with-class-user": {
      "get": {
        "summary": "Get all bookings with class name and user name, sorted by class_datetime",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/bookingdetails-by-user/{id}": {
      "get": {
        "summary": "get booking details by user id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/bookings-by-user/{id}": {
      "get": {
        "summary": "get bookings by user id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/bookings/{id}": {
      "get": {
        "summary": "get a booking by id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "summary": "delete a booking by id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/bookings/": {
      "put": {
        "summary": "update a booking by id",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "example": "any"
                  },
                  "userID": {
                    "example": "any"
                  },
                  "classID": {
                    "example": "any"
                  },
                  "datetime": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}